// frontend/app/app/types/api.ts
export interface components {
    schemas: {
// ...
        CoursePage: {
            /** Items */
            items: components["schemas"]["CourseCard"][];
            /** Total */
            total: number;
            /** Page */
            page: number;
            /** Page Size */
            page_size: number;
            /** Has More */
            has_more: boolean;
            /** Categories */
            categories: string[];
            /** Levels */
            levels: string[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}